requirements.txt
python-dotenv
Pillow
openai
streamlit
